requirements.txt
paho-mqtt==1.6.1
requests==2.30.0
django==3.2.18
django-environ==0.10.0
djangorestframework==3.14.0